Rename settings and dispatcher modules
