+++ tests/test_app.py
@@ -3,8 +3,8 @@ import concurrent.futures
 from aiohttp import web
 
 import virtool.app
-import virtool.app_dispatcher
-import virtool.app_settings
+import virtool.dispatcher
+import virtool.settings
 import virtool.jobs.manager
 
 
@@ -36,7 +36,7 @@ class TestInitSettings:
 
         await virtool.app.init_settings(app)
 
-        assert isinstance(app["settings"], virtool.app_settings.Settings)
+        assert isinstance(app["settings"], virtool.settings.Settings)
 
     async def test_load_called(self, mocker, loop):
         """
