+++ virtool/api/settings.py
@@ -2,7 +2,7 @@ import os
 
 import aiohttp
 
-import virtool.app_settings
+import virtool.settings
 import virtool.http.routes
 import virtool.resources
 import virtool.utils
@@ -16,7 +16,7 @@ async def get(req):
     return json_response(req.app["settings"].data)
 
 
-@routes.patch("/api/settings", admin=True, schema=virtool.app_settings.SCHEMA)
+@routes.patch("/api/settings", admin=True, schema=virtool.settings.SCHEMA)
 async def update(req):
     """
     Update application settings based on request data.
@@ -30,7 +30,7 @@ async def update(req):
 
     settings = req.app["settings"]
 
-    error_message = virtool.app_settings.check_resource_limits(proc, mem, settings.data)
+    error_message = virtool.settings.check_resource_limits(proc, mem, settings.data)
 
     if error_message:
         return conflict(error_message)
@@ -38,7 +38,7 @@ async def update(req):
     proc = proc or settings["proc"]
     mem = mem or settings["mem"]
 
-    error_message = virtool.app_settings.check_task_specific_limits(proc, mem, data)
+    error_message = virtool.settings.check_task_specific_limits(proc, mem, data)
 
     if error_message:
         return conflict(error_message)
